refactor: extract sourcemap utilities
